Flow HS changes for migrating swap endpoints

+++ src-java/testing/functional-tests/src/test/groovy/org/openkilda/functionaltests/spec/flows/FlowHistorySpec.groovy
@@ -2,12 +2,15 @@ package org.openkilda.functionaltests.spec.flows
 
 import static org.openkilda.functionaltests.extension.tags.Tag.LOW_PRIORITY
 import static org.openkilda.functionaltests.extension.tags.Tag.VIRTUAL
+import static org.openkilda.functionaltests.extension.tags.Tag.SMOKE
 import static org.openkilda.functionaltests.helpers.thread.FlowHistoryConstants.CREATE_ACTION
 import static org.openkilda.functionaltests.helpers.thread.FlowHistoryConstants.CREATE_SUCCESS
 import static org.openkilda.functionaltests.helpers.thread.FlowHistoryConstants.UPDATE_ACTION
+import static org.openkilda.functionaltests.helpers.thread.FlowHistoryConstants.UPDATE_SUCCESS
 import static org.openkilda.testing.Constants.NON_EXISTENT_FLOW_ID
 
 import org.openkilda.functionaltests.HealthCheckSpecification
+import org.openkilda.functionaltests.extension.failfast.Tidy
 import org.openkilda.functionaltests.extension.tags.Tags
 import org.openkilda.messaging.payload.history.FlowEventPayload
 import org.openkilda.testing.model.topology.TopologyDefinition.Switch
@@ -15,12 +18,12 @@ import org.openkilda.testing.model.topology.TopologyDefinition.Switch
 import spock.lang.Narrative
 import spock.lang.Shared
 
+import java.util.concurrent.TimeUnit
+
 @Narrative("""Verify that history records are created for the create/update actions.
 History record is created in case the create/update action is completed successfully.""")
 @Tags([LOW_PRIORITY])
 class FlowHistorySpec extends HealthCheckSpecification {
-    String updateHistoryActionV1 = "Updated the flow"
-
     @Shared
     Long timestampBefore
 
@@ -41,6 +44,30 @@ class FlowHistorySpec extends HealthCheckSpecification {
         assert flowHistory.size() == 1
         checkHistoryCreateV1Action(flowHistory[0], flow.id)
 
+        and: "Flow history contains all flow properties in the dump section"
+        with(flowHistory[0].dumps[0]) { dump ->
+            dump.type == "stateAfter"
+            dump.bandwidth == flow.maximumBandwidth
+            dump.ignoreBandwidth == flow.ignoreBandwidth
+            dump.forwardCookie > 0
+            dump.reverseCookie > 0
+            dump.sourceSwitch == flow.source.datapath.toString()
+            dump.destinationSwitch == flow.destination.datapath.toString()
+            dump.sourcePort == flow.source.portNumber
+            dump.destinationPort == flow.destination.portNumber
+            dump.sourceVlan == flow.source.vlanId
+            dump.destinationVlan == flow.destination.vlanId
+            dump.forwardMeterId > 0
+            dump.forwardStatus == "IN_PROGRESS" // issue 3038
+            dump.reverseStatus == "IN_PROGRESS"
+            dump.reverseMeterId > 0
+            //dump.allocateProtectedPath == flow.allocateProtectedPath // issue 3031
+            //dump.encapsulationType == flow.encapsulationType
+            //dump.pinned == flow.pinned
+            //dump.pathComputationStrategy == flow.pathComputationStrategy
+            //dump.periodicPings == flow.periodicPings
+        }
+
         when: "Update the created flow"
         flowHelper.updateFlow(flow.id, flow.tap { it.description = it.description + "updated" })
 
@@ -50,15 +77,63 @@ class FlowHistorySpec extends HealthCheckSpecification {
         assert flowHistory1.size() == 2
         checkHistoryUpdateAction(flowHistory1[1], flow.id)
 
+        while((System.currentTimeSeconds() - timestampAfterUpdate) < 1) {
+            TimeUnit.MILLISECONDS.sleep(100);
+        }
+
         when: "Delete the updated flow"
-        flowHelper.deleteFlow(flow.id)
+        def deleteResponse = flowHelper.deleteFlow(flow.id)
 
         then: "History is still available for the deleted flow"
         def flowHistory3 = northbound.getFlowHistory(flow.id, timestampBefore, timestampAfterUpdate)
         assert flowHistory3.size() == 2
         checkHistoryDeleteAction(flowHistory3, flow.id)
+
+        cleanup:
+        !deleteResponse && flowHelper.deleteFlow(flow.id)
     }
 
+    @Tidy
+    @Tags(SMOKE)
+    def "History records are created for the create/update actions using custom timeline (v2)"() {
+        when: "Create a flow"
+        def (Switch srcSwitch, Switch dstSwitch) = topology.activeSwitches
+        def flow = flowHelper.randomFlow(srcSwitch, dstSwitch)
+        flowHelper.addFlow(flow)
+
+        then: "History record is created"
+        Long timestampAfterCreate = System.currentTimeSeconds()
+        verifyAll(northbound.getFlowHistory(flow.id, timestampBefore, timestampAfterCreate)) { flowH ->
+            flowH.size() == 1
+            checkHistoryCreateV1Action(flowH[0], flow.id)
+        }
+
+        when: "Update the created flow"
+        def flowInfo = northbound.getFlow(flow.id)
+        flowHelper.updateFlow(flowInfo.id, flowInfo.tap { it.description = it.description + "updated" })
+
+        then: "History record is created after updating the flow"
+        Long timestampAfterUpdate = System.currentTimeSeconds()
+        verifyAll(northbound.getFlowHistory(flow.id, timestampBefore, timestampAfterUpdate)){ flowH ->
+            flowH.size() == 2
+            checkHistoryUpdateAction(flowH[1], flow.id)
+        }
+
+        while((System.currentTimeSeconds() - timestampAfterUpdate) < 1) {
+            TimeUnit.MILLISECONDS.sleep(100);
+        }
+
+        when: "Delete the updated flow"
+        def deleteResponse = flowHelper.deleteFlow(flow.id)
+
+        then: "History is still available for the deleted flow"
+        northbound.getFlowHistory(flow.id, timestampBefore, timestampAfterUpdate).size() == 2
+
+        cleanup:
+        !deleteResponse && flowHelper.deleteFlow(flow.id)
+    }
+
+    @Tidy
     def "History records are created for the create/update actions using default timeline"() {
         when: "Create a flow"
         def (Switch srcSwitch, Switch dstSwitch) = topology.activeSwitches
@@ -79,14 +154,18 @@ class FlowHistorySpec extends HealthCheckSpecification {
         checkHistoryUpdateAction(flowHistory1[1], flow.id)
 
         when: "Delete the updated flow"
-        flowHelper.deleteFlow(flow.id)
+        def deleteResponse = flowHelper.deleteFlow(flow.id)
 
         then: "History is still available for the deleted flow"
         def flowHistory3 = northbound.getFlowHistory(flow.id)
-        assert flowHistory3.size() == 2
+        assert flowHistory3.size() == 3
         checkHistoryDeleteAction(flowHistory3, flow.id)
+
+        cleanup:
+        !deleteResponse && flowHelper.deleteFlow(flow.id)
     }
 
+    @Tidy
     def "History should not be returned in case timeline is incorrect (timeBefore > timeAfter)"() {
         when: "Create a flow"
         def (Switch srcSwitch, Switch dstSwitch) = topology.activeSwitches
@@ -105,10 +184,11 @@ class FlowHistorySpec extends HealthCheckSpecification {
         then: "History record is NOT returned"
         flowH.isEmpty()
 
-        and: "Cleanup: restore default state(remove created flow)"
+        cleanup: "Restore default state(remove created flow)"
         flowHelper.deleteFlow(flow.id)
     }
 
+    @Tidy
     def "History should not be returned in case flow was never created"() {
         when: "Try to get history for incorrect flowId"
         def flowHistory = northbound.getFlowHistory(NON_EXISTENT_FLOW_ID)
@@ -125,7 +205,7 @@ class FlowHistorySpec extends HealthCheckSpecification {
 
     void checkHistoryUpdateAction(FlowEventPayload flowHistory, String flowId) {
         assert flowHistory.action == UPDATE_ACTION
-        assert flowHistory.histories.action[-1] == updateHistoryActionV1
+        assert flowHistory.histories.action[-1] == UPDATE_SUCCESS
         checkHistoryCommonStuff(flowHistory, flowId)
     }
